remove previous modal feature.
keep track of which element opened the modal, as there could be multiple triggers for a single modal.

remove partially implemented feature for launching modals within modals.  this is a bad idea and people should feel bad for wanting to do it.

Negate default behavior for links progressively enhanced into buttons.