be wary of initial state
There *could* be a situation where a wrapping element that would get an `inert` and `aria-hidden` applied to it when a dialog opens, actually already has one or both of those attributes.

On dialog close, the script shoudl respect that and put things back the way they were prior to the dialog opening, rather than what it was previously doing, which was just stripping those attributes away.