re-allow escape key to close alert dialog
native alert/confirm dialogs do allow a user to close them via the escape key, so it was overzealous of me to negate that functionality.

Authors should likely map an `esc` close to a “cancel” or “dismiss” button if an alert is asking for a “yes” or “no” sort of input.