Use requestAnimationFrame when focusing dialog
Wrap the `focusTarget.focus()` in a `requestAnimationFrame`. This almost consistently solves for the iOS VO focus issue, where VO focus is not moved to an opened dialog if the dialog has `display: none` as its default display, or if a dialog contains a `role=document` element as its first child, instead of a heading. Credit to Thomas Jaggi, codepen.io/backflip, for recommending the workaround.